requirements.txt
flask==2.2.3
flask-pymongo==2.3.0
pymongo==4.3.3
requests==2.28.2
python-dotenv==1.0.0
openai==0.28.1
schedule==1.1.0
gunicorn==20.1.0
bcrypt==4.0.1
Werkzeug==2.2.3
Flask-Login==0.6.2
email-validator==2.0.0
pytz==2023.3
